Start to decompose integration tests into multiple files
